LPS-124816 Separate in different components
